move Config to individual source file
